Refactored the Headless Gamma API.
