+++ plugins/headless/hu.bme.mit.gamma.api.headless/api/gamma-wrapper.yaml
@@ -56,9 +56,9 @@ paths:
       responses:
         200:
           description: Successfully uploaded the file
-        401:
+        404:
           description: Did not provide a valid workspace
-        403:
+        400:
           description: Project already exists under this workspace, delete it and resend this request
 
   /gamma/api/{workspace}/{projectName}/{filePath}:
@@ -91,7 +91,7 @@ paths:
       responses:
         200:
           description: The operation has started
-        401:
+        404:
           description: Did not provide a valid workspace or project name
         503:
           description: This project is already under operation
@@ -122,7 +122,7 @@ paths:
       responses:
         200:
           description: The operation has stopped
-        401:
+        404:
           description: Did not provide a valid workspace or project name
         503:
           description: This project is not under operation
@@ -159,7 +159,7 @@ paths:
                 properties:
                   filesInWorkspace:
                     type: string
-        '401':
+        '404':
           description: Did not provide a valid workspace or project name
         '503':
           description: This project is not under operation
@@ -204,7 +204,7 @@ paths:
                 format: binary
         503:
           description: The project is currently under operation
-        401:
+        404:
           description: Did not provide a valid workspace or project name
 
   /gamma/deleteproject/{workspace}/{projectName}:
@@ -232,5 +232,58 @@ paths:
           description: Success
         503:
           description: The project is currently under operation
-        401:
+        404:
           description: Did not provide a valid workspace or project name
+  /gamma/status/{workspace}/{projectName}:
+    get:
+      summary: Get the status of the project.
+      operationId: getStatus
+      parameters:
+        - in: path
+          name: projectName
+          required: true
+          description: The name of the project
+          schema:
+            type: string
+            example: gamma.test.project
+        - in: path
+          name: workspace
+          required: true
+          description: The workspace which contains the project
+          schema:
+            type: string
+            format: uuid
+            example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
+      responses:
+        200:
+          description: Success
+          content:
+            application/json:
+              schema:
+                properties:
+                  filesInWorkspace:
+                    type: string
+        404:
+          description: Did not provide a valid workspace or project name
+  '/gamma/addandrun/{workspace}':
+    put:
+      summary: Uploads a project to the workspace and executes a .ggen file
+      operationId: addAndRun
+      parameters:
+        - in: path
+          name: workspace
+          required: true
+          description: The workspace which contains the project
+          schema:
+            type: string
+            format: uuid
+            example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
+      responses:
+        '200':
+          description: Successfully uploaded the project and executed the task
+        503:
+          description: The project is currently under operation
+        404:
+          description: Did not provide a valid workspace or project name
+        400:
+          description: Project already exists under this workspace, delete it and resend this request
