+++ plugins/headless/hu.bme.mit.gamma.api.headless/src/hu/bme/mit/gamma/api/headless/HeadlessApplicationCommandHandler.java
@@ -1,0 +1,17 @@
+package hu.bme.mit.gamma.api.headless;
+
+import org.eclipse.equinox.app.IApplicationContext;
+
+public abstract class HeadlessApplicationCommandHandler {
+	final IApplicationContext context;
+	final String[] appArgs;
+	
+	
+	public HeadlessApplicationCommandHandler(IApplicationContext context, String[] appArgs) {
+		this.context = context;
+		this.appArgs = appArgs;
+	}
+	
+	public void execute() throws Exception {
+	}
+}
